requirements.txt
# Modern Emotion Recognition System - Dependencies

# Core ML and AI libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
scikit-learn>=1.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Web Interface
streamlit>=1.25.0

# Configuration and Utilities
PyYAML>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For better performance
accelerate>=0.20.0
# Optional: For GPU support (uncomment if needed)
# torch-audio>=2.0.0
# torch-vision>=0.15.0

# Legacy support (if needed for migration)
# fer>=22.5.0  # Original FER library - deprecated
